Update patch set 1
Patch Set 1: Code-Review-2

I'm very sorry but I don't think this is how this should be done. That was also one of the reasons why I asked you to split the patch. Thank you very much for following that advice.

1. In general, CSS should not be put in PHP files. CSS belongs in CSS files.

2. This patch makes the selection of images a per-project setting, right? We already have a place where this belongs to: MediaWiki:Common.css. Which is exactly where this setting is now:
https://de.wikipedia.org/wiki/MediaWiki:Common.css (search for ".FA").
Other projects made this a per-skin setting:
https://en.wikipedia.org/wiki/MediaWiki:Vector.css

I once had the oportunity to do a bit of research on this topic:
https://meta.wikimedia.org/wiki/User:TMg/Bugs/40810
My conclusion was to:
* As a first step, add CSS class names (that's the other patch) but no styles.
* Let the communities decide how these classes should look like via their local MediaWiki: CSS files.
* Maybe, if this was discussed and consensus was found, apply default images for all smaller wikis that don't want or can't manage their own MediaWiki: CSS settings. But this (in contrast to this patch) should be a global and not a per-project setting.

I may change my mind if good arguments are given. So please respond if you think I'm wrong.

Patch-set: 1
Label: Code-Review=-2