Update patch set 4
Patch Set 4: Code-Review-1

(1 comment)

I don't think the icons belong in any PHP settings file, see change I818dbb3433b6618c4e1967018765d44f02e9970c. At the moment the values of the array we are discussing here are simply meaningless. The documentation should describe this as it is (and not change the meaning of the array, this will break existing installations). To make it clear the values should be empty strings by default.

What I really would love to see is a setting for meaningful CSS class names:

'badgeClassNames' => array( 'Q987' => 'FA', 'Q654' => 'GA' ),

This should be a per-repo setting that's the same for all clients. The constructed class name should always be added, no matter if a meaningful class name was configured or not. Example: class="badge-Q987 FA".

How these CSS class names are rendered must be defined in a CSS file.

But this should be an other patch, obviously. Please revert this patch back to "add documentation" as it originally was.

Patch-set: 4
Label: Code-Review=-1