Update patch set 2
Patch Set 2:

keeping it to use SimpleSiteLink means more likely we won't forget to change it back to Wikibase\DataModel\SiteLink when someone (probably me, soon) updates usage of SimpleSiteLink -> SiteLink in Wikibase.

it's likely i'll search for all occurrences of "SimpleSiteLink" but not DataModelSiteLink. (although no harm in having the latter, it is just inconsistent)

Patch-set: 2
Label: Verified=0