Update patch set 4
Patch Set 4:

Ok, so finally the discussion seems to concerning client only. As this is a repo setting, we need no values but only the ids to set and thus can remove the key => value structure. Additionally there cannot be any problem with backwards compability as badges are not even fully implemented yet and thus used nowhere.

Patch-set: 4