Update patch set 2
Patch Set 2: Code-Review-1

Thanks for the new patch set! :-) I wonder what's wrong with the test. Looks like Jenkins is broken again. Just two details: Now this looks like it's introducing a setting that's nowhere used. This is at least a bit confusing and probably a mistake. And: I'm not sure if .badge-featured-article { ... } is a good class name. The class names I suggested on IRC where .wb-badge-FA { ... }.

Patch-set: 2
Label: Code-Review=-1