Update patch set 2
Patch Set 2:

Yes, .wb-badge-goodarticle is very good. :-) Originally I wanted to be closer to the original .FA and .GA but there is no point in that if it gets a prefix anyway. So yes, .wb-badge-goodarticle is cool.

I'm sorry that this takes so long but here is one more thing. Please make the option array( 'Q123' => 'wb-badge-goodarticle' ) and not just array( 'Q123' => 'goodarticle' ). The string should be a CSS class name, not just a part of a CSS class name that gets constructed somewhere else. That wouldn't solve my original "don't construct CSS" problem if the CSS class name is still constructed and the .wb-badge- prefix is enforced. K.I.S.S., please.

Patch-set: 2