Update patch set 2
Patch Set 2:

Ok, I get your point that you want to have it as configurable as possible. However, I think the prefix is necessary in this situation to prevent unwanted collisions with might result from a wrong config here. Additionally, we do not construct css here but only classes which is html. ;-)

Patch-set: 2
Label: Verified=0