Update patch set 3
Patch Set 3:

Added the necessary code in the frontend, removed duplicate space, rebased. I'm still not that happy with having such a generic DOM ID claimed, but for this to be useful for linking, it probably must be that way. Also, we should add an anchor so that people actually find out about this feature.

Patch-set: 3
Reviewer: Adrian Heine <1366@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0