Update patch set 3
Patch Set 3:

Wait, I'm afraid I got confused in my first review. This is not about the ItemId, this is about all the PropertyIds. This will work for now but will be a problem if the UI allows multiple Items per page. Are we going to accept this solution for now nevertheless?

Patch-set: 3
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0