Update patch set 2
Patch Set 2:

> the prefix is necessary in this situation to prevent unwanted collisions

There is a prefix in my proposed array( 'Q123' => 'wb-badge-goodarticle' ).

If a bad configuration causes collisions, this is an error in the configuration. It's not our job to fix bad configurations. On the other hand, if a community really wants to stick to the existing class="FA" they should be able to write array( 'Q123' => 'FA' ). It's pointless to have such a configuration in the first place if it still uses enforced prefixes, just like .badge-Q123 does.

Patch-set: 2