Update patch set 4
Patch Set 4: Code-Review+1

I'm curious why the removed method wasn't marked as deprecated? I assume it's not used any more. If that's true (as the tests should have shown) I think there is no reason to not hit +2 on this one. Just want an other pair of eyes to do this.

Patch-set: 4
Label: Code-Review=+1