DNM Insert HTML for badges in ItemView
The javascript edit interface will be added in another patch [1]. The CSS also
needs some adjustments and the box should get an arrow pointing to the star,
like in the mockup [2]. However, there won't be icons for every badge for now
and also no "remove badge" link because we support multiple badges per
sitelink. However, we will make use of the same edit tool which is also used
for aliases and slightly modify it for our requirements (again see [1] for that
purposes). The final mockup for this implementation is [3].

[1] https://gerrit.wikimedia.org/r/#/c/122224/
[2] https://commons.wikimedia.org/wiki/File:Site_link_badges_edit_interface.png
[3] https://commons.wikimedia.org/wiki/File:Site_link_badges_edit_interface_%283%29.png

Bug: 58990
Change-Id: I03a3f6efe6b632cf87df4926d0f50bf5917c4ec9