DNM Insert HTML for badges in ItemView
The star icon is not final but only a first idea [1]. Also there are some
icons missing because I am still waiting for the design team to create
some nice ones. The javascript edit interface will be added in another
patch [2]. The CSS also needs some adjustments and the box should get an
arrow pointing to the star, like in the mockup [3]. However, there won't
be icons for every badge for now and also no "remove badge" link because
we support multiple badges per sitelink. However, we will make use of the
same edit tool which is also used for aliases and slightly modify it for
our requirements (again see [2] for that purposes).

[1] https://commons.wikimedia.org/wiki/File:Five_Pointed_Star_Solid.svg
[2] https://gerrit.wikimedia.org/r/#/c/122224/
[3] https://commons.wikimedia.org/wiki/File:Site_link_badges_edit_interface.png

Bug: 58990
Change-Id: I03a3f6efe6b632cf87df4926d0f50bf5917c4ec9