Update patch set 4
Patch Set 4:

Sorry, I don't think it makes sense to merge this without the code that uses this setting. Maybe it makes more sense to name it "badgeClassNames" or something like that? Maybe there should be a way to do something like

'badges' => array(
	'Q123' => 'wb-badge-goodarticle'
	'Q124' => null
),

to get something like

<a class="badge-Q123 wb-badge-goodarticle">...</a>
<a class="badge-Q124">...</a>

We can't know without having actual code to play around with.

Patch-set: 4