Update patch set 4
Patch Set 4:

> it only contains badges which should be displayed.

Actually, the setting itself does not display anything. All it does is adding CSS class names to some HTML elements.

> the example Q124 with a null value simply would not appear

Why not? If an installation does want to stick with the generated .badge-Q123 class names why should that be forbidden?

Also: If a site link does have multiple badges, I think all class names should be added. Which class is more important should be decided by the order of the classes in the CSS file. Again, we need to see the code to talk about this.

Patch-set: 4