Insert classes for badges on client
Uses the created hook SkinTemplateGetLanguageLink to add a CSS class
for each badge of a sitelink. The setting badgeCssClasses introduced in
another change is also used to determine which badge should be displayed
primarily. For the purpose another css class specified by this setting is
added and the description of this badges' item is used as the title of the
icon indicating the badge using the 'itemtitle' property to set the title
of the <li> element. The icons themselves are also added in another
change.

The clients can add additional styles or override the ones from the config
by using their local Commmon.css page in the MediaWiki namespace. They can
rely on the fact that each badge is always added as a class in the form
"badge-Q123" regardless of whether it is specified in the config or not.

Bug: 60717
Change-Id: If4252b7804ad017075684b794d572843ab6e1ccb