Update patch set 9
Patch Set 9:

From what I see this is just reimplementing the exact same logic that's in SiteLinkTable::getEntityIdForSiteLink.

The real reason why the test fails is that it calls getDBkey which strips the namespace prefix. If you edit the file MockRepository and replace getDBkey with getPrefixedDBkey the testGetSiteLinks test succeeds with the old implementation.

Patch-set: 9