Update patch set 2
Patch Set 2:

As the docs of this class say, and as I have already mentioned when we last discussed this, it is not supposed to be used outside of DM. Hence the internal NS. I'd just leave these occurrences for now rather then trying to make them somewhat nicer. After all, they are not causing a big problem now, and making them somewhat nicer is ultimately not fixing the real problem. However I think we are wasting way to much time on this already, so if you really want to do this, fine. Please do move the class out of the Internal NS first though.

Patch-set: 2