Update patch set 1
Patch Set 1: Code-Review-1

I agree that displayBadges might be a very misleading name. However, siteLinkBadges also does not contain meaningful information as badges always belong to sitelinks. I think we should somehow indicate that the badges listed there are actually badges that appear in the ui by default in opposite to badges not listed there. Maybe we can find a name that is not too confusing but indicates exaclty this point.

Patch-set: 1
Label: Code-Review=-1