Update patch set 1
Patch Set 1:

Sorry for the additional comment but I just realized that "badgeClassNames" also might not fit that well. We have to somehow indicate that not all badges but only the visible ones are listed in that config variable. The best name I can think of is "visibleBadgesClassNames" which is however too long imho. So what do you think about the idea already mentioned above, "visibleBadges"? This makes sure everybody knows why we have the config there and that we need class names there is also mentioned in the comment. Furthermore, "visible" already associates css in its meaning. ;-)

Patch-set: 1