Update patch set 6
Patch Set 6: Code-Review-1

I agree with Lydia.

Page-move archives are an incredibly poor way (in my opinion) to do things. In addition - adding unnecessary features to code is also a poor thing to do. Unless this has any other use case besides as a result of poor archiving methods, I'm a -1.

On a side note - We dealt with this before on-wiki by sysop protecting the item. That is also a poor way to deal with this.

Patch-set: 6
Reviewer: John F. Lewis <589@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1