Update patch set 1
Patch Set 1: Code-Review-1

(3 comments)

Overall this change is very good. Please let me know if I can help you. If it's ok for you I can do some of the fixes I mentioned in the comments and push my patch set on top of yours. But I don't want to confuse you more ;-) so please tell me if it's ok for you if I do this.

Patch-set: 1
Label: Code-Review=-1