Update patch set 5
Patch Set 5:

(1 comment)

I don't think it makes sense to split up the title part from the rest because the title is connected to the css class. Otherwise we could have a situation where the title is added for one badge but the icon shows another badge. To prevent such a inconsistent situation we have to restrict the display of badges.

Patch-set: 5