Render links to deleted items in red
This is intended to be a quick intermediate step on the way to the
propper change I9bbbe8055d68b274fa6141eecff4a8d9414f89d9.

This patch does not introduce additional lookups to the database.
So it shouldn't have any unwanted side-effects. All it does is
rendering links to deleted items in red instead of blue. This should
fix bug 61854.

Would be cool to have this merged before the next sprint. :-)

Bug: 61854
Change-Id: Iec03211105a921669e140b0d3cbbabeec4caed97