Update patch set 2
Patch Set 2:

(2 comments)

This patch is not supposed to be a perfect, final solution. It's a step on the way. It makes the code more flexible and following patches a lot easier to create and review.

The reset() calls are not supposed to stay. They will be replaced with proper formatters in following patches. As I said in the description.

Please help me moving forward and not get stuck for another month.

Patch-set: 2
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0