Update patch set 16
Patch Set 16:

This keeps confusing me. This removes all localization. Why? Is this supposed to stay unlocalized? Or do we want this localized and this patch is just a quick hack because the current solution does not work well and "always English" is better than sticking with a broken solution?

Patch-set: 16