Rename formatter option RESOLVE_ID to LOOKUP_LABEL
This is a follow-up to one of the comments in
Iec03211105a921669e140b0d3cbbabeec4caed97.

I'm going introduce an EntityTermLookup in a future patch. It will
be used to lookup labels, descriptions and if the Entity exists.
What the option does is turning this "label lookup" on/off.
It doesn't necesarry "resolve an ID".

Again, this patch alone may look a bit pointless. It's part of a
much bigger change.

Change-Id: If065f9fe39ccfd50c715f6d5a8a39bdf3189708d