Stop throwing fatal errors in CommonsLinkFormatter
We are currently using the Title object as a passive formatter to
normalize file names (e.g. replace underscores with spaces). This
may not be the best solution but it works. We may think about other
ways to do this but as long as we are using this method we should
not throw wild "Fatal error: Call to a member function [...] on a
non-object" to the frontend and logs.

Bug: 62904
Change-Id: I28a20dbb63a8eb353b80de4e055e0eb830695efe