Update patch set 1
Patch Set 1: Code-Review-1

This changes the code so it does not throw an exception if JobQueueGroup::push() returns a non-bool value.

But it always returned a bool as far as I can see (I checked the current master and the history). I don't think this can fix the current problem we have.

-1 because I think it's correct to throw an exception if $ok is null.

Patch-set: 1
Label: Code-Review=-1