Factor out generation of site links HTML into own class
This has the nice side effect of reducing the number of created SiteSQLStore
instances. Since every SiteSQLStore instance maintains its own cached version
of the sites list, which it first has to fetch from the DB or cache, this
reduces both rendering time and memory consumption.

SectionEditLinkGenerator needs tests and
SectionEditLinkGenerator::getHtmlForSiteLinkGroup should be split up for that.

Change-Id: Id2e1fba574b737c15fa2dc5ee4fb6152ca240c96