[WIP] Introduce EntityContent::STATUS_UNTITLED
This is a direct follow-up to change
I72d671712c000cce2831770895e06626c0fa6f51.

ItemContent currently contains a bit of dead code. The case
$status === self::STATUS_EMPTY && $hasSiteLinks
can never happen because Item::isEmpty already considers the
site links and sets the status to stub, not empty.

This change removes this dead code and introduces a new status.
I think this new status can be very helpful (more helpful than
the LINKSTUB status) but I'm not sure if the team agrees.
Therefor the WIP.

Change-Id: I9523656af1effab76965572c4382e863b88c1d0c