Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

> resolved in a clean way (removing the node)

In my opinion hiding the node is the most clean way. That's the main idea of this patch, to reduce the possible reflows/redraws to a single CSS class that's added/removed in a single step.

The spinner is an empty div with two classes, that's all. Whoever finds it there, he will be able to understand why it's there. I don't see the problem of leaving it in the DOM. Removing it by manipulating the DOM again can cause more problems (e.g. performance wise) than leaving it where it is.

I'm thinking about a possible follow-up that outputs the div via PHP. Then all that's left to do in JS is adding/removing the class. Which can very easily be done in vanilla JS.

Patch-set: 2
Label: Code-Review=-1