Update patch set 5
Patch Set 5:

Generally looks good, though I think the whole entity status calculation is misplaced. "entity status calculation" already hints at this: it's about the status of an Entity. So why is this in EntityContent?

This is not something I'd like to see in Entity itself, or rather it's derivatives, either. Having a method such as isLinkStub in Item would be fine however, and that could then be used by the status code.

How about creating an object for figuring out the status for an Item and one for a Property? Those could then be delegated to from a method in ItemContent and PropertyContent, respectively. That method would be an implementation of an abstract one in EntityContent.

Patch-set: 5