Update patch set 8
Patch Set 8: Code-Review-1

I want to re-think when and where knowledge about the concrete entity type is brought into the information flow. Perhaps it would be better to just use the info we have in Changeop::validate( Entity ) anyway - the ChangeOp would then need a ValidatorProvider that can provide the appropriate validator(s) for a given type of entity to be used by the ChangeOp.

Alternatively, the entity type could be passed to all the newXxxOp functions in the ChangeOpfactories, but that seems reven more annyoing than the current situation.

Patch-set: 8
Label: Code-Review=-1