Update patch set 4
Patch Set 4:

I played around with PHP's DateTime and found a few things.

I tried to detect if DateTime found a time zone or not. And let the parser fail if a time zone was found. On the other hand, why should we disallow the input of time zones? This doesn't make sense and is not really the problem we have.

The problem is that PHP's parser sets the time, day and month to the current if you use it to parse incomplete dates. "1980" becomes "1980-04-29T01:10:31+0200". Wow. Who designed this parser? There is no way to disable this stupid "feature" or to detect if it kicked in or not. Really, I tried so many things. Screw it.

My solution: Make the DateTime parser require at least three numbers.

if ( !preg_match( '/\d+\D+\d+\D+\d/', $string ) ) {
	throw Exception( 'This parser needs at least three numbers.' );
}

This may still lead to wrong results in some cases but it blocks all cases we know at the moment: a single year, year and month (there are other parsers for these) and a year with a timezone (e.g. "1980+3").

Patch-set: 4