Update patch set 5
Patch Set 5:

@thiemo I agree generally about requiring three parts to the value, so we have month, date and year.

DateTime indeed behaves not the way we want if we have less.

If we do have just year, though it should be handled by YearTimeParser, and month-year by YearMonthTimeParser, and '1980s' or '1980er' should be handled by MWTimeIsoParser. If it's none of the above (such as these edge cases with year + timezone), then TimeParser will attempt with this DateTimeParser and then weirdness happens.

I'm trying to cover these edge cases best as possible, though would be happy with follow up to do that further better than this.

Patch-set: 5