Update patch set 11
Patch Set 11:

Are you sure this diff does not contain other patches that are already in an other Gerrit review? I see so much code here I already commented or even merged in previous patches.

It would be really helpful to do the namespace changes in an other patch. If that's possible.

Patch-set: 11
Label: Verified=0