Update patch set 22
Patch Set 22: Code-Review-1

I think this is doing the right thing in the wrong place. Calling a TimeFormatter on e.g. a StringValue is a programming error and should trigger an InvalidArgumentException. It should not happen. I suggest to handle this further up the stack. That would be clearer and require fewer changes to the code.

The issue at hand arises when rendering a *Snak* that contains a DataValue of a different type than what is expected based on the Snak's Peroperty's data type. That's bug 63299; Please tran and fix that and see if any problems remain (diff may need special handling, but that should also be done on the level of Snaks).

As I wrote in my comment to the bug, this problem should be avoided by:

* Get data type for property
* Get expected data value type for property data type
* If expected data value type does not match actual data value type, add error to output (optionally; this should be a FormatterOption). The formatting of the error must depend on the expected output format, of course (plain text vs. html - wikitext can probably use the HTML version).
* look up formatter based on property data type if expected data value type matches actual data value type
* If no formatter found (or the value type didn't match), look up formatter for actual data value type

Patch-set: 22
Reviewer: Daniel Kinzler <128@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1