Fix viewing older Entity revisions
This is done by constructing an EntityRevision object
with the old revision id instead of just using the current
(latest one) in EntityContent.
After having fixed that, we can use this EntityRevision
object to fetch the correct version of the Entity for
the deferred rendering (EntityViewPlaceholderExpander)
of the "In other languages" box.

Bug: 63309
Change-Id: Id6fbbc85148d5efe6545d5a91a99b5ec19e96665