Reduce expensive tests in MwTimeIsoFormatter
As far as I can tell these 397 (for all languages) extremely
expensive tests (they need several minutes on my machine) never
tested much. They are here to make sure that numbers in languages
with a digitTransformTable are returned with normalized/unlocalized
digits. To succesfully test this it's enough to test a few
languages that are known to have transform tables.

Change-Id: I888728ced2cdff5bb7888ce54402c271e6f705e3