Update patch set 1
Patch Set 1:

Even 49 seconds is way to much for such a simple thing like a time formatter. The problem is that this initializes all ~400 languages. The Language code executed is explicitly marked as expensive. The reason for the extreme run time on my machine is that I have most caches turned off.

As explained I'm sure it's simply not necessary to test all languages. It's not even worth 49 seconds. Test a few languages that are known for the problem that must be tested.

Patch-set: 1
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0