Remove .wbc-editpage in css
.wb-langlinks-link is present everywhere now and is sufficient
as a selector for adding the pencil.

.wbc-editpage additionally exists as a separate link (to wikidata!,
since the item is connected, but has no site links except enwiki).
This causes double pencil.

Iac189e7 does empty() on the parent element of the links, so
the link and extra pencil goes away, but might as well not
have double pencil and legacy css.

one pencil remains on span .wb-langlinks-link in all cases.

Bug: 65001
Change-Id: Ib96e6aa8c16cf6f6db9dd121bf3a0afcffa9e285