Update patch set 2
Patch Set 2: Code-Review-1

(9 comments)

Not passing the "ignore conflicts" array around is a good idea. It's so much easier to read now. And really not worth the trouble. It's one or two elements in an array that usually contains one or two elements.

I think this is ready to be merged if the comments are addressed. Just a few copy & paste errors and naming suggestions.

Patch-set: 2
Label: Code-Review=-1