Update patch set 16
Patch Set 16:

The ValueValidator interface is generic, it's not defined to apply to any specific kind of subject type. So, using it for strings is not abuse. I'm not adding implementations of that interface, I'm using existing implementation and apply them to more things in our data model (labels, descriptions and aliases).

I'm happy to discuss the fact that the ValueValidator interface is overly broad, and think about ways to remedy that. But at the moment, it's not obvious to me how to do that. Adding another interface with exactly the same signature for validating strings does not seem useful.

So, let's please go on with this patch, and then talk about improving the validation of data values separately.

Patch-set: 16