Attempt to clean-up and refactor ChangeOp and Validators code
This is not really fixing a bug. The code works. But it's hard to
read and to maintain. And it's error-prone since some versions of
the data model return new objects when calling getFingerprint()
but the recent versions don't. This is a problem because this code
actually relies on getFingerprint() returning a new object.
Unfortunatelly the only solution I have so far is to misuse the
scary serialize functions. I really don't like these in production
code.

This is a direct follow up to some (but not all) of the comments in
Ibb6ee8439ba3ab0252403eb811aa48d1274d1d58.

Change-Id: I7434d5c40cd601976e72f104ce5699e59fef184e