Attempt to clean-up and refactor ChangeOp and Validators code
This is not really fixing a bug. The code works. But it's hard to
read and to maintain. Unfortunatelly the best I could do is to use
clone. I tried to avoid it. It gets a mess very fast (try to
reconstruct a Fingerprint object from an other Fingerprint object).
This just doesn't feel right. I think we really need to discuss this.

This is a direct follow up to some (but not all) of the comments in
Ibb6ee8439ba3ab0252403eb811aa48d1274d1d58.

Change-Id: I7434d5c40cd601976e72f104ce5699e59fef184e