Use deep cloning in ChangeOpLabel/Description/Aliases
The clone keyword in PHP doesn't really (deep) clone. But this is
really important in certain versions of the data model.

Plus a bunch of refactoring. I found this code hard to read and to
maintain.

This is a direct follow up to some (but not all) of the comments in
Ibb6ee84.

Change-Id: I7434d5c40cd601976e72f104ce5699e59fef184e