Right-to-left rendering for site links in diff
Example: https://www.wikidata.org/wiki/Special:Diff/131600143

I know this code probably needs a lot more love. But I did not
wanted to waste to much time at the moment. The solution works and
does not make the code worse, in my opinion. It's the same
combination of attributes as in the site links table, see
$templates['wb-sitelink'].

Bug: 36635
Change-Id: Ic29f287c6be423cda6d945a84628fc24ddc8a9ed