Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

Looks good for me, so it's almost a +1 from my side except for a tiny little detail (see comment). But I would like to hear what the others (Aude, Jeroen, Daniel) have to say.

A question, Bene*: Could it be that ClientSiteLinkLookup is currently not used? Where will it be used?

General question: An other way to do the same is to put this caching directly into SiteLinkTable? What do you guys think about this idea? The fingerprint of this cache of ItemId object would be very small (ItemId only haves 2 properties). But the impact could be high because SiteLinkTable::getItemIdForLink is used almost 20 times in API, Lua and special pages.

Patch-set: 1
Label: Code-Review=-1