Update patch set 5
Patch Set 5: Code-Review+1

I think the removal of the original "if labels are equal in both languages, don't show both" should not be part of this patch. But since the design of the special page is bad anyway this doesn't make it worse. It's a bit more consistent. Always show the second language if your search and interface languages are different.

I will merge it as it is if you still want.

Patch-set: 5
Label: Code-Review=+1