Update patch set 1
Patch Set 1: Code-Review+2

This change is consistent with what we have generally been doing. The test base class, be it the MW one or the PHPUnit one, is in the top level NS, in which case it is somewhat silly to import it if it is only uses in the extends clause. We can of course discuss if we want to change this, though it makes little sense to block this change based on that.

Patch-set: 1
Label: Code-Review=+2