Update patch set 1
Patch Set 1: Code-Review-2

(1 comment)

Please don't remove @var and constructor docs just because the tools you use are able to find the information somewhere else. I'm fine if you don't want to add this to "your" code. But please don't remove this from existing code. It makes the code unnecessary hard to read if you have to check the constructor first to find out what the type is. Not every reader uses an IDE or code analysis tool just to read and understand code. A common way to look at our code is the web interface. Which does not do any IDE magic.

Having these docs everywhere does have several benefits. The information is at the same place in every file. No matter what tool somebody uses he is able to find the information within seconds. It will never look like somebody forgot the doc if it's non-optional. It's part of the style guide like the controversial spaces inside brackets are. They may be meaningless but we all use them for a simple reason: to make the code easier to read.

And please try to split this into more than one patch.

Patch-set: 1