Remove second/strict check for edit conflicts.
Edit conflicts are usually handled by EditEntity, but there exists
a race condition for the case that a concurrent edit gets saved between
the time EditEntity checks and the time the new revision is saved.

This patch removes a workaround we used to mitigate that race condition.

Noting that core does not do a second check for regular wikitext
edits and just accepts the unlikely case of a concurrent edit getting
saved during the critical split second, there is no reason for Wikibase
to employ such a nasty workaround to achieve that a check.

If a strict check for edit conflicts based on an "anchor" revision is
desired, this should be implemented in core.

Note that an earlier attempt to rely on the $baseRevId for a late check
does not work, since core uses $baseRevId to indicate the "logical
parent" revision, i.e. the revision the user loaded for editing, which
is different from the the current revision at the time of editing e.g.
in the case of a rollback. This would cause all rollbacks to fail with
an edit conflict.

Change-Id: I4a33cff85ba88eebc5cfe669b038aed1700c68e1