Update patch set 2
Patch Set 2:

If you -1 because you think it can be improved you should tell how you think it can be improved.

The list is created by Status::getWikiText(). The test makes sure the HTML list elements are not escaped. It fails with the old code and succeeds with your patch. Which is all we need to know.

I added a second assertion that makes sure the output does not contain any double escaped characters. But as Daniel pointed out I can not make this a positive test.

If you know how to inject a dummy message please let me know.

Patch-set: 2
Label: Verified=0