Update patch set 2
Patch Set 2: Code-Review+1

Switching to protected is a very good idea.

Not sure about the parameters. The original idea obviously was to be able to override the methods in the subclasses and have access to all three revision numbers. Which was never used but may be in the future.

Compatibility with core is not the point. These methods aren't used in other code.

Patch-set: 2
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1