Update patch set 3
Patch Set 3: Code-Review-1

You could check the html output you get against a parsed version of the error message you expect. Right now this does test that html is not escaped, but that wasn't the point of the change you want to unit test here... the point was to have wikitext parsed.

Patch-set: 3