Stick to ContentHandler::getDefaultFormat default implementation
Overriding the existing ContentHandler::getDefaultFormat with a
conflicting implementation that calls an other getDefaultFormat
implementation that really does nothing else than returning a
constant (which will never change according to the documentation)
feels like an unnecesarry deep Inception to me.

ContentHandler already knows how to pick a default format. No need
to override this.

The idea "this should never change" belongs to a test. Not sure if
my test is the best we can have to ensure this but it does the job.

Change-Id: Ibc4e5122cff776550df9d650fa887d8dbe7a021e