Expand tooltip details fast
There is no reason to slow this down so much. The collapsed "details"
part usually contains a very short one or two line message. The
visual distraction is very minor. A smooth animation is probably nice
but should be as fast as possible. This is just a text error message
after all.

Change-Id: Ia4711c930d6c6f1801217767b6ccf77250e562de