Don't use services in EntityContent
Content objects shoudl ideally be value objects. This moves
any code that relies on services into the associated ContentHandler,
instead of having it in the Content object itself. This allows
such services to be injected when constructing the handler.

Note that the construction of EntityView objects is moved to
EntityHandler for now, but should be factoried into a separate
factory eventually.

Change-Id: I99753f806fa37bd21cc122a68dd47c4565484b25