Update patch set 2
Patch Set 2: Code-Review-2

Does this mean the first and the last shard contain unbalanced brackets? That's invalid JSON. I understand the idea (you are looking at the shards as strings that need to be concatenated, not as individual JSON blobs). But shouldn't it be possible to parse the shards independently?

Patch-set: 2
Label: Code-Review=-2