Update patch set 1
Patch Set 1:

You wrote: "Inject it via the constructor and set it in JS or so... that's at least a valid way for inter-component data exchange". Sorry, but that's pointless. Why should I set CSS via JavaScript if there is a CSS file for exactly that reason? It's a client-only CSS file that overrides some existing styles from existing components. That's exactly what the CSS cascade is made for. Moving the exact same logic to JavaScript doesn't make this any better. The opposite happens. It's more confusing because nobody expects CSS buried in JavaScript.

If you can find a way to detect the dynamic z-index from the dialog via JavaScript, fine, go ahead and submit a patch that does this. In any case, this patch belongs to the client and not to the ValueView component.

Patch-set: 1