Update patch set 6
Patch Set 6:

Oh, and the shards in PS6 will never be valid JSON. Not before and not after concatenation. When we discussed this in IRC my impression was that the solution should be more or less what your PS1 did (add the opening "[" to the first shard and the closing "]" to the last shard) but make this behavior an option. Keep the current behavior (each shard is valid JSON) when the option is not set.

I may be wrong. Please explain.

Patch-set: 6