Update patch set 2
Patch Set 2: Code-Review+2

Because it's a subclass of ApiTestCase which is a subclass of MediaWikiLangTestCase. There are other ways to solve this, see I623c2924a207aeaec7777e833decf5bf9a17d2b9, but I'm not sure if this is applicable in this case.

Patch-set: 2
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+2