Cleanup EntityView, SectionEditLinkGenerator and related
Main change is that the edit URLs (they are required for
non-JavaScript browsers) can be null if there is no entity id.
Even if it is possible to call the special pages without an entity
id this does not make sense if you are clicking a link in an entity
view that does not have an id.

The reason to touch this code was what I saw in
Id2e1fba574b737c15fa2dc5ee4fb6152ca240c96.

Change-Id: I8f3f029b59dfdd23ad5561fcb41e303463b3a296