Make LanguageFallbackChain compatible to EntityInfoBuilder
This is split from my pretty old draft I9bbbe80. It expands the
"language chain lookup" (or fixes a bug by adding a missing feature
that is already described in the documentation of the class).

Short explanation: If the provided data array is not an array of
strings but an array of arrays (like the arrays returned by the
EntityInfoBuilders) just return the array as it is.

Additional test cases cover this behavior.

Change-Id: I715bbb02aa780451d4431d185261b3a648910312