Update patch set 1
Patch Set 1: Code-Review-1

Several notes:

This script needs to be able to operate of a *large* list of IDs, read from a file. It could optionally also operate on all pages in a given namespace. Giving a small list on the command line does not seem useful - a small set op pages can be touched manually just as well.

Note that, when doing a large number of database updates, we need to periodically wait for slaves to catch up. 

@Hoo: A null edit can not change the serialization. The relevant line in WikiPage is:
   $content->equals( $old_content )

equals() is not based on the serialized form, but on the internal representation (an Entity object, in the case or Wikibase). 

As to filtering the DataUpdates: we need some way to filter thjem, but class name no longer works, since we now use DataUpdateClosure for all of them. Perhaps DataUpdateClosure could have a getName() method which we could match against.

Regarding "re-parsing later": that means posting jobs to the job queue. I think RefreshLinks is the right one, but I'm not sure.

Patch-set: 1
Reviewer: Daniel Kinzler <128@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1