Update patch set 7
Patch Set 7:

Combining different dumps can't work that way. You will never end with a valid JSON blob. The first one needs an opening "[", all but the last need a "," at the end and the last one needs a "]" but no ",". How would you do that?

It seems we are talking about different things. How the option is implemented now it really _is_ about the format of the shards. Controlling the format of different requests would be an other option.

Patch-set: 7