Update patch set 7
Patch Set 7: Code-Review-1

Seems I totally misunderstood Daniel. His idea is to keep this option as simple as possible:

a) Default is as it is now. Each shard is valid JSON with opening and closing brackets.
b) No shards contains brackets. No shard ends with a comma.

If you implement it like that then Daniel is right: This is not a sharding option. It's useful in all cases, also with sharding turned off. You can combine sharded and non-sharded results from different queries. You can ignore the order of the shards completely. These two things are clear advantages, in my opinion. The downside is that you need to add the commas and brackets by yourself if you want to combine stuff. But that's acceptable, isn't it?

Very sorry about the confusion. All blame on me, please.

Patch-set: 7
Label: Code-Review=-1