Update patch set 3
Patch Set 3:

(6 comments)

Sorry. I agree that the old ChangeNotifier class was bad. But this new approach is not better in my opinion. The most silly details from the old class are not touched. This only changes an implementation detail that's more a matter of taste than anything else.

Please try to understand what _concept_ you want to wrap in a class and find a name and structure that describes that concept.

Ending up with a "dummy channel" is clearly wrong. You can have a channel with nobody listening, yes. You can have a sender without a receiver, yes. You can have a notifier but nobody who registered to be notified, yes.

Unfortunately none of this describes the current implementation. In it's core this thing currently calls a save() method thats defined in the IORMRow interface and implemented in the ORMRow class. This is neither "channeling" nor "sending" not "notifying", this is "doing".

Better keep the old implementation then.

Patch-set: 3