Update patch set 3
Patch Set 3:

After a chat with Jeroen, what about this: I will merge this along with the next patch if it's ok for you when I do a follow-up that completely removes this class. So we will introduce it as an intermediate solution and almost immediately drop it. Deal?

Patch-set: 3