Update patch set 3
Patch Set 3:

(5 comments)

Thiemo: if you want to get rid of this class completely, please explain how you plan to encapsulate the knowledge of *how* notifications are sent to client wikis. Writing a follow up would be a good way to explain :)

Saving a row to the database *is* not sending. Neither is pushing an object on a queue. Or calling an interrupt. These things may, in the right context, *implement* sending a message.

Patch-set: 3