Make DataFieldValidator require the field to be set.
DataFieldValidator should fail validation when the respective field
is not set (missing or null). Or current use of DataFieldValidator is
consistent with that assumption.

Previously, DataFieldValidator would fail hard if the field as unset,
and passed the value to the sub-validator if the field was null, which
was not consistent with the current usage.

Change-Id: I64c139c11f60b69f7a7a674bb95c87934ce58861