Update patch set 3
Patch Set 3:

@Thiemo: I did. My question stands. We need an abstract for "send this notification to client wikis". We need another one for "stuff happened, make a notification (and send it off)". These should be separate, hence the naming rochade. 

Yes, you could just have an optional ChangesTable in the "outer" make-a-notification-for-stuff object. But that entangles two concerns instead of separating them: the thing that makes notifications will then have to know how to send them, making it harder to mock, and to replace with a different implementation.

Patch-set: 3